code review changes extracted delete dialog
